requirements.txt

PySide6~=6.7.2
requests~=2.31.0
mysql-connector-python~=9.0.0
arcgis~=2.3.0